Update tb.v
Verilog code of the testbench used to test the final BIST circuit. It displays the value of the sequence and sign at each step and then displays the Final Signature at the end. The signature calculated from the implementation here matches the value of the signature in the handwritten notes, depicting the successful operation of the final BIST circuit.